pyproject.toml
[tool.poetry]
name = "lotion-api"
version = "0.9.0"
description = ""
authors = ["ask4git <ask4git@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2.16"
djangorestframework = "3.15.2"
psycopg2 = "^2.9.10"
gunicorn = "^23.0.0"
shortid = "^0.1.2"
drf-rw-serializers = "^1.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
